backend/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
python-multipart==0.0.6
boto3==1.34.5
openai-whisper==20231117
moviepy==1.0.3
celery==5.3.6
redis==5.0.1